# libs/ts/identity-sdk-core/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "nvbes Identity API",
    "description": "Authentication, authorization, and identity management API",
    "contact": {
      "name": "nvbes",
      "url": "https://nvbes.fr"
    },
    "license": {
      "name": "UNLICENSED"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "https://identity.nvbes.fr",
      "description": "Production"
    },
    {
      "url": "http://localhost:8080",
      "description": "Development"
    }
  ],
  "paths": {
...
    "/auth/challenge/pwd": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "challenge_pwd",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PwdRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResult"
                }
              }
            }
          },
          "202": {
            "description": "Password accepted, MFA challenge required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentifierResult"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials or state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/tenants/{tenantId}/domains": {
      "get": {
        "tags": [
          "federation"
        ],
        "operationId": "list_domains",
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "description": "Tenant ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List tenant domains",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantDomainsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "federation"
        ],
        "operationId": "create_domain",
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "description": "Tenant ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTenantDomainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Domain created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantDomainResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/tenants/{tenantId}/domains/{domainId}": {
      "delete": {
        "tags": [
          "federation"
        ],
        "operationId": "delete_domain",
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "description": "Tenant ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "domainId",
            "in": "path",
            "description": "Domain ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Domain deleted"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateTenantDomainRequest": {
        "type": "object",
        "required": [
          "domain"
        ],
        "properties": {
          "domain": {
            "type": "string"
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "request_id": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
      "IdentifierResult": {
        "type": "object",
        "required": [
          "next_step",
          "state_token"
        ],
        "properties": {
          "available_methods": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "next_step": {
            "type": "string"
          },
          "state_token": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "LoginResult": {
        "type": "object",
        "required": [
          "user",
          "session"
        ],
        "properties": {
          "session": {
            "$ref": "#/components/schemas/SessionView"
          },
          "user": {
            "$ref": "#/components/schemas/UserView"
          },
          "verification_resend_available_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
...
      "PwdRequest": {
        "type": "object",
        "required": [
          "state_token",
          "password"
        ],
        "properties": {
          "password": {
            "type": "string"
          },
          "state_token": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "SessionView": {
        "type": "object",
        "required": [
          "id",
          "created_at",
          "last_seen_at",
          "expires_at",
          "current"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "current": {
            "type": "boolean"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "ip": {
            "type": [
              "string",
              "null"
            ]
          },
          "last_seen_at": {
            "type": "string",
            "format": "date-time"
          },
          "organization_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "revoked_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "tenant_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "user_agent": {
            "type": [
              "string",
              "null"
            ]
          },
          "workspace_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "workspace_region": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "TenantDomainResponse": {
        "type": "object",
        "required": [
          "domain"
        ],
        "properties": {
          "domain": {
            "$ref": "#/components/schemas/TenantDomainView"
          },
          "verification_token": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "TenantDomainView": {
        "type": "object",
        "required": [
          "id",
          "domain",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "domain": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "verification_expires_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "verification_requested_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "verified_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
      "TenantDomainsResponse": {
        "type": "object",
        "required": [
          "domains"
        ],
        "properties": {
          "domains": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TenantDomainView"
            }
          }
        }
      },
...
      "UserView": {
        "type": "object",
        "required": [
          "id",
          "email",
          "display_name",
          "email_verified",
          "mfa_enabled",
          "created_at"
        ],
        "properties": {
          "birthdate": {
            "type": [
              "string",
              "null"
            ],
            "format": "date"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "display_name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "email_verified": {
            "type": "boolean"
          },
          "firstname": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "lastname": {
            "type": [
              "string",
              "null"
            ]
          },
          "mfa_enabled": {
            "type": "boolean"
          },
          "region": {
            "type": [
              "string",
              "null"
            ]
          },
          "username": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    }
  },
...
}